pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "bookstore API"
authors = [
    {name = "marcelo chaves",email = "marceloluiz.icg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
